% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoMerge.R
\name{QN}
\alias{QN}
\title{Quantile normalization for multiple vectors of possibly different sizes}
\usage{
QN(lov)
}
\arguments{
\item{lov}{a list of numeric vectors}
}
\description{
Quantile normalization for multiple vectors of possibly different sizes
}
\note{
Uses linear interpolation if the sizes are different
}
