% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoMerge.R
\name{split_matrix}
\alias{split_matrix}
\title{splitting a matrix to a list by a margin according to a vector of groups}
\usage{
split_matrix(mat, vector, byrow = TRUE)
}
\description{
splitting a matrix to a list by a margin according to a vector of groups
}
