% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoMerge.R
\name{QN.xp}
\alias{QN.xp}
\title{Quantile normalization on a matrix + vector of groups (splitting by rows)}
\usage{
QN.xp(xp_mat, groups)
}
\description{
Quantile normalization on a matrix + vector of groups (splitting by rows)
}
